requirements.txt
# Core dependencies
websockets==14.1
python-dotenv==1.0.1
loguru==0.7.2

# LLM Providers
langchain-community>=0.4.0
langchain-openai>=1.1.0
langchain-anthropic>=1.4.0
google-genai>=0.4.0

# LLM Client libraries
cohere==5.21.1
openai>=2.0.0
anthropic>=0.96.0

# Speech-to-Text
deepgram-sdk==3.11.0

# Data handling
pydantic==2.10.5

# Development dependencies
pytest>=8.3.0
pytest-asyncio>=0.25.0
black>=24.0.0
flake8>=7.1.0
